roll_detector/requirements.txt
# 基础依赖
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO11 (需要本地安装)
# pip install -e /path/to/yolo11

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 评估指标
scikit-learn>=1.3.0

# 其他
tqdm>=4.65.0
